+++ src/z80/test/cpu.stack.test.js
@@ -1,8 +1,8 @@
 import expect from 'expect.js';
-import CPU from '../src/cpu.js';
-import Memory from '../src/memory.js';
+import CPU from '../cpu.js';
+import Memory from '../memory.js';
 // import { random8bit } from './utils.js';
-import rom from '../roms/flappyboy.json';
+import rom from '../../../roms/flappyboy.json';
 
 describe('CPU Stack', () => {
   let cpu, memory;
